requirements.txt
#TODO brænd ubrugte libs og funktioner af :)

astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.4.4
pyserial==3.4
requests==2.23.0
six==1.14.0
typed-ast==1.4.1
urllib3==1.25.8
wrapt==1.11.2
